Fixes related to JDK8 in SamFileHeaderMergerTest
1. SAM test files had spaces as delimiters in
   some headers (spec requires tab delimiters)
2. The order of attributes in merged headers
   is non-deterministic (and turns out to be
   different in jdk8 than in earlier versions).
   This tests the headers independent of
   attribute order.
3. jdk8 tearDown complains when tmpDir is empty
   in SortingCollectionTest, earlier jdks
   ignored this.

+++ src/tests/java/htsjdk/samtools/util/SortingCollectionTest.java
@@ -55,10 +55,12 @@ public class SortingCollectionTest {
 
     @AfterTest void tearDown() {
         System.err.println("In SortingCollectionTest.tearDown.  tmpDir: " + tmpDir);
-        for (final File f : tmpDir.listFiles()) {
-            f.delete();
+        if (tmpDir.exists()) {
+            for (final File f : tmpDir.listFiles()) {
+                f.delete();
+            }
+            tmpDir.delete();
         }
-        tmpDir.delete();
     }
 
     protected boolean tmpDirIsEmpty() {
